requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
dj-database-url==0.5.0
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.6.2
django-cors-headers==3.4.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
lml==0.0.9
Pillow==8.0.1
psycopg2==2.8.5
pyexcel-io==0.5.20
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.1.0
xlrd==1.2.0
